+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Bug Fix
 - `Include missing` is now always displayed on the design report
+- Saving numeric and integers with domains no longer forces the domain to cover all the captured values
 
 ## 0.14.4 (March 20, 2013)